DEVREL-955 Move calloutType into a separate file
