+++ lib/types/perCollection.ts
@@ -1,5 +1,3 @@
-import { contentTypes } from "../../models";
-
 export type PerCollection<T> = Readonly<{
   ficto_healthtech: T;
   ficto_imaging: T;
@@ -16,8 +14,3 @@ const emptyCodenames: PerCollection<null> = {
   ficto_surgical: null,
   ficto_healthtech: null,
 };
-
-export type CalloutType = keyof typeof contentTypes.callout.elements.type.options;
-
-export const isCalloutType = (codename: string | undefined): codename is CalloutType => 
-  !!codename && Object.keys(contentTypes.callout.elements.type.options).includes(codename);
